rename Trainworks.Base.Enum to Trainworks.Base.Enums C# doesn't like namespace == name of Class.
