+++ TrainworksReloaded.Base/Enums/CharacterTriggerTypeFinalizer.cs
@@ -11,7 +11,7 @@ using TrainworksReloaded.Core.Extensions;
 using TrainworksReloaded.Core.Interfaces;
 using UnityEngine;
 
-namespace TrainworksReloaded.Base.Enum
+namespace TrainworksReloaded.Base.Enums
 {
     public class CharacterTriggerTypeFinalizer : IDataFinalizer
     {
